Move validation methods to utilities object
